+++ NEWS.md
@@ -11,6 +11,7 @@
     - Minor adjustments for edge cases in which summary variables are `NULL`
     - Added clarification to table captions
     - Eliminated irrelevant error and warning message printouts from rendered report
+    - Remove inconsequential "bad bags" from Agenda tab; do not report 1s received instead of 0s, or 0s received instead of 1s
 - Edited `identicalBags()` function to exclude matching coots_snipe and rails_gallinules from MI in output; this state uses the response from one question to populate both fields.
 
 # migbirdHIP 1.2.8