+++ NEWS.md
@@ -17,6 +17,7 @@
     - Edited "bad bags" section of agenda tab to return a message in the edge case of 0% of a file containing bad bag values
     - Edited the Issuance tab to reflect new rules in evaluating if a record is current
 - Edited `identicalBags()` function to exclude matching coots_snipe and rails_gallinules from MI in output; this state uses the response from one question to populate both fields.
+- Edited `read_hip()` function to exclude "hold" subdirectories when reading season HIP data.
 
 # migbirdHIP 1.2.8